Make global config version optional
Without this commit, the global "version" setting needs to always be set even if
it is explicitly set in every pe_bootstrap provisioner instance. This
also means that the provisioner level config that inherits the global
will fail if a global is set and no provisioner level config is specified.
If there is a global value, reasonably the provisioner should be able to
omit it.

This commit removes the failure when "version" is unset. It leaves the
responsibility of verifying a global "version" is present with the
provisioner configuration that needs it.