Use merged config in the provisioner
Without this commit, there is no simple way to chose values from
global/local scope for use in the provisioner class. This is currently
manifesting as an inability to specify global values and then use locals
to override them.

This commit makes use of the new merge functionality in the
PEBuild::Util::Config module. This commit merges the values from the two
configurations and then uses the resulting object in the provisioner. It
also passes the configuration to the "download" and "unpackage" actions
since they hjave no access to global configurations.