Make global and provisioner configs work together
Without this commit, the configuration classes do not work together
coherently. The provisioner configuration inherits the global
configuration and along the way calls many of the parent class
methods. The validate method in particular is returns a hash that is hard
coded to the class its originating from. While the provisioner class
should fall back on the parent class for validation of inherited variables,
the resulting error messages need to be provided in the context of the
provisioner connfiguration. Also, the provisioner does not currently
validate that it has a value for the "version" variable in the event one
is not set globally. This should be a failure.

This commit adds a validation for missing "version" value on both global
and provisioner configurations. This commit also takes any errors
originating from the parent class and adds puts them in context of the
provisioner configuration before passing them along.