Don't use parent "finalize" in provisioner config
Without this commit the parent finalize method is called from the
PEBuild::Config::PEBootstrap class. This causes the
PEBuild::Config::PEBootstrap class to establish defaults for the
variables inherited from the global class. When this happens, merging is
not effective. In this case it is more functional to not have any
defaults for the inherited values and assume they will be merged in. Any
values that do not have at least a default in PEBuild::Config::Global
should have validation performed.

This commit removes the call to the parent finalize method in
PEBuild::Config::PEBootstrap.