Archive prints available installers on error
Without this commit, the archiver does not try to fail when no method of
finding an installer is encountered. Instead it simple just waits to
fail on the download attempt.

This commit changes the archiver to detect when an installer is not
present and also when no download_root is available. The archiver will
now respond with a message letting the user know no installer of the
desired version is available and display the list of available
installers.