Add I18n Support
Without this commit, there is no I18n support in pe-build. I18n is used
in various places in Vagrangt including exceptions classes. Without this
support, a new exception class would require overriding a fair amount of
functionality in the `Vagrant::Errors::VagrantError` class. There are
also various other places this support will be beneficial as well.

This commit adds I18n support to this plugin. This is achieved by adding
a load statement and a basic default translation file for the english
language.