Use exception to fail in PEBuild::Archive
Without this commit, the `PEBuild::Archive` class would fail out of
execution using the `Vagrant::Errors::VagrantError` exception. This is
not very specific or handleable for other systems relying on or working
with the PEBuild plugin.

This commit adds a `PEBuild::ArchiveNoInstallerSource` exception class
and the appropriate message to the english translation file for the I18n
component of Vagrant execption classes.