Change default filename param to include :version
Without this commit, the default value used by the global configuration
does not properly handle version if it is not defined globally. There is
already functionality in `PEBuild::Archive` that will replace the
string ":version" with the version number passed from the merged
configurations.

This commit updates the default value for the global configuration
parameter "filename" to use the string ":version".