Lazy binding of variables for global 'filename' default
Without this commit, the default value of the config 'filename'
attirbute was dependant on the values of the @version and @suffix
variables pre-merge time. Since these variables need only be set at
either global or local levels, there is no way to resolve their usage in
this default until the merge.

This code changes the default value to use the @version and @suffix
variables of the local object (which will be post merge in this case) in
the event the @filename variable is set to
`Vagrant::Plugin::V2::Config::UNSET_VALUE`.