Trap SSLError and SocketError during download
During `vagrant up` (specifically, during the `import` action), if an
exception is raised that does not inherit from `VagrantError` then
Vagrant will forcibly destroy the box.

This patch rescues two additional exception classes raised during download of
PE installers:

  - `SocketError` which can be raised if an incorrect URL is provided.

  - `SSLError` which can be raised if a `https` URL cannot be validated.

The errors are re-raised as `DownloadFailed` which is a subclass of
`VagrantError`.