Add Windows Server 2003 support
Windows Server 2003 support wasn't added as I didn't originally have a
test machine or a need. After comments in Github and a few more spare
minutes it seemed like an easy win, so here it THEORETICALLY is. This is
not yet tested on a Windows Server 2003 machine, but it IS tested on
Windows Server 2008 (works great) and SHOULD also work on Server 2003.