[config/pe_bootstrap] Normalize role value
Symbols are preferred over strings when referencing hash keys and
discrete string-ish values in Ruby, but they're a major pain to
specify in configuration files. For instance, the syntax for specifying
symbols in YAML is `!ruby/sym role`, which is very clumsy.

Instead of forcing users to handle this little idiosyncrasy, this commit
interns the `role` value when specified so that end users don't have to
know about this, and consuming code doesn't have to normalize strings
and symbols.