Intelligently set ext for posix pre-release builds
The detect_installer capability of the pe_bootstrap provisioner will try
to intelligently determine the PE installer filename for a given
platform. On Posix, PE releases are packaged as .tar.gz files. However,
if you have access to pre-release builds of PE, those builds will not be
gzipped, they will simply be .tar files.

This commit updates the posix detect_installer capability to understand
when a pre-release build is being asked for, and set the extension to
"tar" in that scenario. For released builds, the "tar.gz" extension is
appropriately selected.