Fix regression in merging version numbers
This patch removes the `nil` default from `Config::PEBoostrap` so that it
doesn't overwrite a version set in `Config::Global`. As a consequence of this,
an unset version can no longer be assumed to default to `nil` unless
`Util::Config.local_merge` is used to merge in a global config.

Fixes #68.