Add abstract capability for getting facts from VMs
This commit adds a new guest capability, `pebuild_facts`, which loads facts
from guest VMs. If Facter is installed, it will be used to generate a full list
of facts. If Facter is not installed, a basic set of info is computed about the
`os` and `architecture`, which should be sufficient to drive a PE installation.