Update acceptance tests for pe_agent
This patch updates the acceptance suite to provision CentOS PE 2015.x masters
with `pe_bootstrap` and Ubuntu 2015.x agents with `pe_agent`. Acceptance setup
is also expanded to include support for both CentOS and Ubuntu agent platforms.

Because of the expanded platform support, other tests are also updated to
reflect the change in Vagrant box names.