Purge agent when VMs are destroyed
The `pe_agent` provisioner can now be configured with `autopurge` set to a true
or false value. When enabled, the `pe_agent` provisioner will purge agent
certificates and data from the master when the agent VM is destroyed.  This
setting defaults to `true` if the `master_vm` setting has been set.

Closes oscar-stack/vagrant-pe_build#54