Stop pe_agent from mutating config.master_vm
Prior to this patch, the `pe_agent` provisioner would replace the `String`
value of `config.master_vm` with an actual `Vagrant::Machine` instance. This
causes an error if the provisioner configuration is re-validated after the
provisioner runs. It also screws up the simplicity of the data model for
pe_agent configuration.

This patch stores the `Vagrant::Machine` reference as an instance variable on
the provisioner instead of mutating the configuration.

Fixes oscar-stack/vagrant-pe_build/issues/100