(GH-105) Fix agent purge when master_vm is missing
Prior to this commit, the `cleanup` action of the `pe_agent` provisioner would
attempt to use the `master_vm` when warning that cleanup was being skipped
because the master wasn't created. This caused an error since, `master_vm` was
nil. This patch sends the message using the agent instead.

Fixes oscar-stack/vagrant-pe_build#105