Raise exception if bad certname value is passed
Previous to this commit, if a bad value for certname somehow made it
passed validation, the
PEBuild::Provisioner::PEAgent#provsion_posix_agent method would simply
accept it, causing the default behavior for the
pe_frictionless_installer.sh script to be used.  While not catastrophic,
this could cause considerable confusion.

This commit throws an exception if a bad value makes it passed
validation. It also prints output showing what value of certname is
being used, if the value can be predetermined.