Add fallback to master_vm if the hostname is not set on master_vm
Prior to this commit, the master setting of the pe_agent provisioner
would not get set by default if the master_vm did not have a
hostname set on it.

After this commit, the master setting will default to the master_vm
setting if the hostname is not set on the master_vm.