Close PE archive files after download
PE Build uses Ruby's OpenURI module to download installer archives. This module
downloads content, places it in a temp file, and then returns an open
filehandle. The filehandle is subsequently used to move the downloaded file
into the PE archive directory. However, move operations fail on windows if the
handle is left open.

This patch ensures the handle is closed prior to the move operation.

Fixes oscar-stack/vagrant-pe_build#60.