Add shared_installer option
This patch adds a new global configuration option, shared_installer. When set
to the default value of true, PE installation is run using resources shared
through the `/vagrant` mount. When set to false, PE installer resources are
downloaded directly to the VM --- which is more efficient for cloud services
and VPN connections.