Set default pe_bootstrap role in provisioner
This patch moves the default for the pe_bootstrap role from the configuration
class to the provisioner class. This allows us to set different defaults based
on version number. This operation must be done in the provisioner as that is
the only place where global version information is available.

For PE versions prior to 2016.2.0, the default of :agent is used. Post
2016.2.0, a new default of :master is used as PE tarballs no longer support
agent-only installs.