Deprecate pe_boostrap agent installs
Starting with PE 2015.2.0, support for running an agent-only installation from
a PE tarball was removed. This operation would still work with some hacks.
However, with 2016.2.0, these hacks no longer work and it is time to officially
remove the ability to run agent-only installs from PE tarballs.

This patch adds deprecation notices to agent installs from 2015.2.0 onwards and
an error for installs under 2016.2.0.

Closes oscar-stack/vagrant-pe_build#95.