Stop staging installers with root privilages
This patch sets the `;sudo` flag to `false` when installer tarballs are
staged and extracted on POSIX VMs. This prevents permission errors that
occur when a non-root account is being used by Vagrant ssh, the unpacked
installer ends up owned by root, and additional files need to be
uploaded to the unpacked installer via scp.

Fixes oscar-stack/vagrant-pe_build#126.