Update GPG key insertion to account for PE layout
Most versions of PE, from 3.0.0 to 2016.1.2, stored the release key in a `gpg/`
subdirectory. However, both 2016.2.0 and 2.x used different paths. Packages
were not signed prior to 2.7. This patch updates the code that swaps in an
extended GPG key to account for the differences in versions.

Fixes oscar-stack/vagrant-pe_build#135