(SUP-698) Add agent type for provisioning HA Replicas
Prior to this commit there were no way to provision a replica
with the `pe_agent` provisioner, which requires commands be run on the
master while provisioning the replica. This commit adds the
capabilitity to specify an `agent_type` as a `replica`, which will run
the HA provisioning on the master.