Expand download path used by System.Net.WebClient
Something recently changed in the way Vagrant uses WinRM that has caused the
System.Net.WebClient to download files to `C:/Windows/System32` instead of
a working directory under a user's home folder when a path such as `./` is
specified. This commit fixes the issue by:

  - Using Get-Item to expand the destination path before passing it to
    the downloader.

The path where the download is saved is also logged to aid in future debugging
expeditions.

Fixes oscar-stack/vagrant-pe_build#139.